style(authAndAccount): separate Auth and Account controllers
