+++ api/controllers/v1/AccountController.js
@@ -1,0 +1,9 @@
+/**
+ */
+
+const accountController = require('../AccountController');
+
+module.exports = {
+  changePassword: (req, res) => accountController.changePassword(req, res),
+  forgotPassword: (req, res) => accountController.forgotPassword(req, res),
+};
