+++ config/routes.js
@@ -91,10 +91,12 @@ module.exports.routes = {
   //  ╠═╣╠═╝║  ║╣ ║║║ ║║╠═╝║ ║║║║║ ║ ╚═╗
   //  ╩ ╩╩  ╩  ╚═╝╝╚╝═╩╝╩  ╚═╝╩╝╚╝ ╩ ╚═╝
 
+  /* Account controller */
+  'PATCH /api/v1/account/password': 'v1/Account.changePassword',
+  'POST /api/v1/forgotPassword': 'v1/Account.forgotPassword',
+
   /* Auth controller */
   'GET /api/v1/logout': 'v1/Auth.logout',
-  'PATCH /api/v1/account/password': 'v1/Auth.changePassword',
-  'POST /api/v1/forgotPassword': 'v1/Auth.forgotPassword',
   'POST /api/v1/login': 'v1/Auth.login',
   'POST /api/v1/signup': 'v1/Auth.signUp',
